SuperStreetViewRouter

Documentation for SuperStreetViewRouter.

SuperStreetViewRouter.benchmark_random_walk_solutions

SuperStreetViewRouter.create_input_graph

SuperStreetViewRouter.generate_output_file

SuperStreetViewRouter.get_city_street

SuperStreetViewRouter.get_total_city_cost

SuperStreetViewRouter.get_total_city_length

SuperStreetViewRouter.solve_graph_greedy

SuperStreetViewRouter.benchmark_random_walk_solutions — Function

benchmark_random_walk_solution(test_iterations::Int64=10)

Generates the provided number of random walk solutions and displays statistics on the lengths of the results. Intended to provide a lower bound upon which to improve.

SuperStreetViewRouter.create_input_graph — Method

create_input_graph(city::City)

Returns a SimpleWeightedDiGraph representing the provided City object. Graph vertices are Junction indices (as stored in the City object) and edges are street lengths.

SuperStreetViewRouter.generate_output_file — Function

generate_output_file(solution::Solution, path=nothing)

Creates a text file using the provided Solution object and saves it to the specified path.

SuperStreetViewRouter.get_city_street — Method

get_city_street(city::City, start_junction::Int64, end_junction::Int64)

Returns a Street object in the provided city that allows travel from the provided start junction index to the provided end junction index. Returns nothing if no such Street is found.

SuperStreetViewRouter.get_total_city_cost — Method

get_total_city_cost(city::City)

Returns the total time required to traverse all streets of the provided city in seconds.

SuperStreetViewRouter.get_total_city_length — Method

get_total_city_length(city::City)

Returns the total length of all streets in the provided city in meters.

SuperStreetViewRouter.solve_graph_greedy — Function

solve_graph_greedy(city::City=read_city())

Generates a greedy solution to the provided city, or uses the default city if none is provided. The greedy algorithm can be described as follows: - At each junction, the car will choose to traverse the street with the highest "value" = length / time. - Traversed streets will be recorded. - Traversed streets will have their length decreased by a constant factor to discourage but allow usage. This constant is likely responsive to optimization.

Inefficient due to only routing a single path at a time. Implementing a shared list of traversed streets and dedicating a separate thread to each path would likely improve performance.