SuperStreetViewRouter

Documentation for SuperStreetViewRouter.

SuperStreetViewRouter.CityData

SuperStreetViewRouter.CityGraph

SuperStreetViewRouter.StreetData

SuperStreetViewRouter.benchmark_random_walk_solutions

SuperStreetViewRouter.create_input_graph

SuperStreetViewRouter.find_best_street

SuperStreetViewRouter.generate_output_file

SuperStreetViewRouter.get_city_street

SuperStreetViewRouter.get_possible_streets

SuperStreetViewRouter.get_total_city_cost

SuperStreetViewRouter.get_total_city_length

SuperStreetViewRouter.solve_graph_greedy

SuperStreetViewRouter.CityData — Type

CityData(total_duration::Int, nb_cars::Int, starting_junction::Int)

A struct that holds the metadata for a city. Has constructor(s)

CityData(c::City)

SuperStreetViewRouter.CityGraph — Type

CityGraph(data::CityData, graph::ValOutDiGraph)

A graph representation of a City object. The graph is a SimpleValueGraphs.jl ValOutDiGraph, and the data is a CityData object.

SuperStreetViewRouter.StreetData — Type

StreetData(duration::Int, value::Int, id::Int)

Structure storing the data required for greedy algorithm in edges of city graph. Has constructors

StreetData(street::Street)
StreetData(street::Street, id::Int)

SuperStreetViewRouter.benchmark_random_walk_solutions — Function

benchmark_random_walk_solution(test_iterations::Int64=10)

Generates the provided number of random walk solutions and displays statistics on the lengths of the results. Intended to provide a lower bound upon which to improve.

SuperStreetViewRouter.create_input_graph — Method

create_input_graph(city::City)

Returns a SimpleWeightedDiGraph representing the provided City object in a directed graph. Graph vertices are Junction indices (as stored in the City object) and edges are weighted by street lengths.

SuperStreetViewRouter.find_best_street — Method

find_best_street(possible_streets)

Returns the best street to traverse from the provided list of possible streets. The best street is the one with the highest value, or the one with the highest value * elapsedstreetpenalty ^ times traversed if it has been traversed before.

SuperStreetViewRouter.generate_output_file — Function

generate_output_file(solution::Solution, path=nothing)

Creates a text file (.txt) using the provided Solution object and saves it to the specified path.

SuperStreetViewRouter.get_city_street — Method

get_city_street(city::City, start_junction::Int64, end_junction::Int64)

Returns a Street object in the provided City that allows travel from the provided start junction index to the provided end junction index whether one-directional or bi-directional. Returns nothing if no such Street is found.

SuperStreetViewRouter.get_possible_streets — Method

get_possible_streets(city_graph)

Returns a list of all streets that can be traversed from the provided junction. The streets are tuples of (endjunction, streetdata)

SuperStreetViewRouter.get_total_city_cost — Method

get_total_city_cost(city::City)

Returns the total time (in seconds) required to traverse all streets of the provided city.

SuperStreetViewRouter.get_total_city_length — Method

get_total_city_length(city::City)

Returns the total length (in meters) of all streets in the provided city.

SuperStreetViewRouter.solve_graph_greedy — Function

solve_graph_greedy(city::City=read_city())

Generates a greedy solution to the provided city, or uses the default city if none is provided. The greedy algorithm can be described as follows: - At each junction, the car will choose to traverse the street with the highest "value" = length / time. - Traversed streets will be recorded. - Traversed streets will have their length decreased by a constant factor to discourage but allow usage. This constant is likely responsive to optimization.

Inefficient due to only routing a single path at a time. Implementing a shared list of traversed streets and dedicating a separate thread to each path would likely improve performance.