SuperStreetViewRouter

Documentation for SuperStreetViewRouter.

SuperStreetViewRouter.CityData

SuperStreetViewRouter.CityGraph

SuperStreetViewRouter.StreetData

SuperStreetViewRouter.apply_penalty

SuperStreetViewRouter.benchmark_random_walk_solutions

SuperStreetViewRouter.create_input_graph

SuperStreetViewRouter.find_best_path

SuperStreetViewRouter.find_best_street

SuperStreetViewRouter.get_path_value

SuperStreetViewRouter.get_possible_paths

SuperStreetViewRouter.get_possible_streets

SuperStreetViewRouter.get_total_city_cost

SuperStreetViewRouter.get_total_city_length

SuperStreetViewRouter.path_time

SuperStreetViewRouter.solve_graph_greedy

SuperStreetViewRouter.CityData — Type

CityData(total_duration::Int, nb_cars::Int, starting_junction::Int)

A struct that holds the metadata for a city. Has constructor(s)

CityData(c::City)

SuperStreetViewRouter.CityGraph — Type

CityGraph(data::CityData, graph::ValOutDiGraph)

A graph representation of a City object. The graph is a SimpleValueGraphs.jl ValOutDiGraph, and the data is a CityData object.

SuperStreetViewRouter.StreetData — Type

StreetData(duration::Int, value::Int, id::Int)

Structure storing the data required for greedy algorithm in edges of city graph. Has constructors

StreetData(street::Street)
StreetData(street::Street, id::Int)

SuperStreetViewRouter.apply_penalty — Method

apply_penalty(pre_penalty_value::Float64, num_traversals::Int64, elapsed_street_penalty::Float64)::Float64

Applies a traversal penalty to a provided value.

SuperStreetViewRouter.benchmark_random_walk_solutions — Function

benchmark_random_walk_solution(test_iterations::Int64=10)

Generates the provided number of random walk solutions and displays statistics on the lengths of the results. Intended to provide a lower bound upon which to improve.

SuperStreetViewRouter.create_input_graph — Method

create_input_graph(city::City)

Returns a SimpleWeightedDiGraph representing the provided City object in a directed graph. Graph vertices are Junction indices (as stored in the City object) and edges are weighted by street lengths.

SuperStreetViewRouter.find_best_path — Method

find_best_path(possible_paths, traversed_streets, elapsed_street_penalty)

Returns the best path to traverse from the provided list of possible paths. The best path is the one with the highest value considering the number of times each street has been traversed with the provided penalty

SuperStreetViewRouter.find_best_street — Method

find_best_street(possible_streets, traversed_streets, elapsed_street_penalty)

Returns the best street to traverse from the provided list of possible streets. The best street is the one with the highest value, or the one with the highest value * elapsedstreetpenalty ^ times traversed if it has been traversed before.

SuperStreetViewRouter.get_path_value — Method

get_path_value(path, traversed_streets, elapsed_street_penalty)

Returns the value of the provided path, taking into account the number of times each street has been traversed with the provided penalty

SuperStreetViewRouter.get_possible_paths — Method

    get_possible_paths(city_graph, current_junction, remaining_time, depth)

returns list of tuples

SuperStreetViewRouter.get_possible_streets — Method

get_possible_streets(city_graph)

Returns a list of all streets that can be traversed from the provided junction. The streets are tuples of (endjunction, streetdata)

SuperStreetViewRouter.get_total_city_cost — Method

get_total_city_cost(city::City)

Returns the total time (in seconds) required to traverse all streets of the provided city.

SuperStreetViewRouter.get_total_city_length — Method

get_total_city_length(city::City)

Returns the total length (in meters) of all streets in the provided city.

SuperStreetViewRouter.path_time — Method

path_time(path)

Returns the time it takes to traverse the provided path

SuperStreetViewRouter.solve_graph_greedy — Function

solve_graph_greedy(city::City=read_city())

Generates a greedy solution to the provided city, or uses the default city if none is provided. The greedy algorithm can be described as follows: - At each junction, the car will choose to traverse the street with the highest "value" = length / time. - Traversed streets will be recorded. - Traversed streets will have their length decreased by a constant factor to discourage but allow usage. This constant is likely responsive to optimization.

Inefficient due to only routing a single path at a time. Implementing a shared list of traversed streets and dedicating a separate thread to each path would likely improve performance.